Rename agent modules to snake_case
