+++ protocols/agents/harmony_synthesizer_agent.py
